[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14176 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -1,16 +1,19 @@
+
+
 import java.io.*;
 import java.util.Scanner;
 import java.util.StringTokenizer;
 
 public class Main {
-    public static int f(int a, int b) {
+    public static long f(long a, long b) {
         return (a + b) * (a - b);
     }
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
-        System.out.println(f(Integer.parseInt(st.nextToken()) , Integer.parseInt(st.nextToken())));
+        System.out.println(f(Long.parseLong(st.nextToken()) , Long.parseLong(st.nextToken())));
 
     }
 }
+
